refactor(window-splitter): reorganize context hook and update exports
- Moved the `useWindowSplitterContext` hook to a new file for better structure.
- Updated exports in `window-splitter.tsx` to include subcomponents with clearer naming.
- Removed unused context-related code from `window-splitter.context.tsx`.
