+++ packages/nimbus/src/components/window-splitter/hooks/use-window-splitter-context.tsx
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { WindowSplitterContext } from "../components/window-splitter.context";
+
+export const useWindowSplitterContext = () => {
+  const context = useContext(WindowSplitterContext);
+  if (!context) {
+    throw new Error(
+      "useWindowSplitterContext must be used within a WindowSplitter.Root"
+    );
+  }
+  return context;
+};
